Fix for collecting files from the transpilation tmp folder
